feat(cli): start improving CLI
