+++ cli/src/cli/commands/mod.rs
@@ -1,0 +1,5 @@
+mod node;
+pub(crate) use node::*;
+
+mod sub_chain;
+pub(crate) use sub_chain::*;
