+++ cli/src/main.rs
@@ -13,5 +13,5 @@ fn main() -> Result<()> {
     let (config, partial_log_config) = seda_config::create_and_load_or_load_config();
     // We hold the guards so logging works properly.
     let _guard = seda_logger::init(&partial_log_config.to_config(options.log_options.clone())?);
-    options.handle(config)
+    options.command.handle(config)
 }
